requirements.txt
PyYAML>=6.0
pywin32>=306
